Fix Sprint Scene Merge
Manually merged the Sprint Scene.
After performing the merge, David's version was the end result.  I copied
the GameObjects he added to the clipboard, then replaced the scene with
the enlargeDrag version and pasted the objects back in.  This was
sufficient to resolve the merge.

We should probably make sure to not modify the same GameObjects or Prefabs
to make merging easier.